Refactor: 회원가입, 로그인, 회원정보 조회, 알림, 프로필 이미지 등록 등 멤버 도메인 기능 예외처리 추가 및 코드 컨벤션에 맞춰 수정
