+++ src/main/java/com/example/demo/global/config/SecurityConfig.java
@@ -62,7 +62,12 @@ public class SecurityConfig {
 
                 // api 경로
                 .authorizeHttpRequests(authorizeHttpRequests -> authorizeHttpRequests
-                        .requestMatchers("/api/hello", "api/members", "api/authenticate", "api/login", "/api/kakao", "/api/product/category/**", "api/product/detail/**", "api/sendSMS/{phone}", "api/chat", "api/latestChat", "api/chat/update", "api/chat/create", "/ws/**", "/app/sendMessage", "/topic/**", "/api/search", "/api/checkName/**").permitAll()
+                        .requestMatchers("/api/hello", "api/signup", "api/authenticate",
+                                "api/login", "/api/kakao", "/api/product/category/**",
+                                "api/product/detail/**", "api/sendSMS/{phone}", "api/chat",
+                                "api/latestChat", "api/chat/update", "api/chat/create",
+                                "/ws/**", "/app/sendMessage", "/topic/**", "/api/search",
+                                "/api/checkName/**").permitAll()
 
                         .anyRequest().authenticated() // 나머지 경로는 jwt 인증
                 )
