+++ src/main/java/com/example/demo/global/security/TokenProvider.java
@@ -46,7 +46,6 @@ public class TokenProvider {
                 .map(GrantedAuthority::getAuthority)
                 .collect(Collectors.joining(","));
 
-        System.out.println(authentication.getName());
         long now = (new Date()).getTime();
         Date validity = new Date(now + this.tokenValidityInMilliseconds);
         return Jwts.builder()
