+++ src/main/java/com/example/demo/src/member/dto/ResponseMyInfo.java
@@ -12,7 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ResponseMyInfo {
     private String email; // 이메일
-    private String password; // 비밀번호
     private String userImage; // 유저 프로필 이미지
     private String name; // 이름
     private String career; // 멘토일 경우 커리어
@@ -25,7 +24,6 @@ public class ResponseMyInfo {
         if (member == null) return null;
         return ResponseMyInfo.builder()
                 .email(member.getUsername())
-                .password(member.getPassword())
                 .userImage(member.getUserImage())
                 .name(member.getName())
                 .career(member.getCareer())
